backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
apscheduler==3.10.4
python-telegram-bot==21.10
openai==1.59.6
python-dateutil==2.9.0.post0
httpx==0.28.1
pytest==8.3.4
